moved navbar and footer
