+++ src/App.jsx
@@ -5,8 +5,9 @@ import AOS from "aos";
 import "aos/dist/aos.css";
 import emailjs from '@emailjs/browser';
 
+import Navbar from './components/layout/Navbar';
+import Footer from './components/layout/Footer';
 import logo from "./assets/logo.png";
-import navLogo from "./assets/logo512.png";
 import solution1 from "./assets/solution1.png";
 import solution2 from "./assets/solution2.png";
 import solution3 from "./assets/solution3.png";
@@ -49,19 +50,23 @@ const contactFormInputStyles = {
 };
 
 function App() {
-  const [isMobileMenuOpen, setMobileMenuOpen] = useState(false);
-  const [setCurrentSection] = useState("home");
-  const [eiSplit, setEiSplit] = useState(true);
-  const [isTechSectionVisible, setIsTechSectionVisible] = useState(false);
   const [isDarkMode, setIsDarkMode] = useState(false);
   const [isLoading, setIsLoading] = useState(true);
-  const [lastScrollY, setLastScrollY] = useState(0);
   const [isMobile, setIsMobile] = useState(window.innerWidth < 768);
+  const [eiSplit, setEiSplit] = useState(true);
+  const [isTechSectionVisible, setIsTechSectionVisible] = useState(false);
   const techSectionRef = useRef(null);
-  const [setTechSectionScrollProgress] = useState(0);
   const [activeCard, setActiveCard] = useState(null);
   const [formStatus, setFormStatus] = useState({ loading: false, success: false, error: null });
   const [formData, setFormData] = useState({ name: '', email: '', message: '' });
+
+  useEffect(() => {
+    AOS.init({
+      duration: 1000,
+      once: true,
+      mirror: false,
+    });
+  }, []);
 
   // Add window resize listener
   useEffect(() => {
@@ -78,11 +83,6 @@ function App() {
       setIsLoading(false);
     }, 2500);
     return () => clearTimeout(timer);
-  }, []);
-
-  // initialize AOS (scroll animations)
-  useEffect(() => {
-    AOS.init({ duration: 800, once: true });
   }, []);
 
   // Intersection Observer for Technology section
@@ -102,57 +102,6 @@ function App() {
     if (techSection) observer.observe(techSection);
     return () => techSection && observer.unobserve(techSection);
   }, []); // Remove scrollDirection dependency
-
-  // update currentSection based on scroll position
-  useEffect(() => {
-    const sections = ["home", "solutions", "technology", "resources", "faqs", "contact"];
-    const onScroll = () => {
-      const scrollY = window.scrollY + 120;
-      let found = "home";
-      for (let id of sections) {
-        const el = document.getElementById(id);
-        if (el && el.offsetTop <= scrollY) found = id;
-      }
-      setCurrentSection(found);
-    };
-    window.addEventListener("scroll", onScroll);
-    return () => window.removeEventListener("scroll", onScroll);
-  },);
-
-  // lock body scroll when mobile menu is open
-  useEffect(() => {
-    document.body.style.overflow = isMobileMenuOpen ? "hidden" : "";
-  }, [isMobileMenuOpen]);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      const currentScrollY = window.scrollY;
-      setLastScrollY(currentScrollY);
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, [lastScrollY]);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      if (techSectionRef.current) {
-        const { top, height } = techSectionRef.current.getBoundingClientRect();
-        const progress = Math.min(1, Math.max(0, (window.innerHeight - top) / height));
-        setTechSectionScrollProgress(progress);
-
-        // Combine letters when scrolled past 50% of the section
-        if (progress > 0.5) {
-          setEiSplit(false);
-        } else {
-          setEiSplit(true);
-        }
-      }
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  },);
 
   // Initialize EmailJS
   useEffect(() => {
@@ -237,15 +186,6 @@ function App() {
     }
   };
 
-  const navLinks = [
-    { id: "home", label: "Home" },
-    { id: "solutions", label: "Solutions" },
-    { id: "technology", label: "Technology" },
-    { id: "resources", label: "Resources" },
-    { id: "faqs", label: "FAQs" },
-    { id: "contact", label: "Contact" }
-  ];
-
   return (
     <div
       className={`font-sans transition-colors duration-300 overflow-x-hidden relative ${
@@ -258,127 +198,7 @@ function App() {
       </div>
 
       {/* NAVBAR */}
-      <nav
-        className={`fixed top-0 left-0 right-0 z-50 border-b transition-colors duration-300 ${
-          isDarkMode ? "bg-gray-800 border-gray-700" : "bg-white border-orange-500"
-        }`}
-        data-aos="fade-down"
-      >
-        {/* max‑width wrapper keeps content centred & equal padding */}
-        <div className="max-w-7xl mx-auto flex items-center justify-between px-4 sm:px-6 lg:px-8 py-2">
-          {/* --- Logo --- */}
-          <img
-            src={navLogo}
-            alt="Enable Intelligence Logo"
-            className="h-8 sm:h-10 w-auto object-contain block"
-            style={{ maxWidth: "120px" }}
-          />
-
-          {/* --- Desktop Nav Links + Dark Toggle --- */}
-          <div className="hidden sm:flex items-center space-x-4">
-            {navLinks.map(link => (
-              <a
-                key={link.id}
-                href={`#${link.id}`}
-                className={`px-3 py-1 rounded-lg border transition ${
-                  isDarkMode
-                    ? "border-gray-600 hover:bg-gray-700"
-                    : "border-black hover:bg-black hover:text-white"
-                }`}
-              >
-                {link.label}
-              </a>
-            ))}
-            {/* Dark‑mode button */}
-            <button
-              onClick={() => setIsDarkMode(!isDarkMode)}
-              className={`p-2 rounded-full transition-all duration-300 ${
-                isDarkMode ? "bg-gray-700 hover:bg-gray-600" : "bg-gray-100 hover:bg-gray-200"
-              }`}
-              aria-label={isDarkMode ? "Switch to light mode" : "Switch to dark mode"}
-            >
-              {isDarkMode ? (
-                <svg className="w-6 h-6 text-yellow-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z" />
-                </svg>
-              ) : (
-                <svg className="w-6 h-6 text-gray-700" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z" />
-                </svg>
-              )}
-            </button>
-          </div>
-
-          {/* --- Mobile: dark toggle + hamburger --- */}
-          <div className="sm:hidden flex items-center space-x-1">
-            {/* Dark toggle */}
-            <button
-              onClick={() => setIsDarkMode(!isDarkMode)}
-              className={`p-1.5 rounded-full transition-all duration-300 ${
-                isDarkMode ? "bg-gray-700 hover:bg-gray-600" : "bg-gray-100 hover:bg-gray-200"
-              }`}
-              aria-label={isDarkMode ? "Switch to light mode" : "Switch to dark mode"}
-            >
-              {isDarkMode ? (
-                <svg className="w-5 h-5 text-yellow-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z" />
-                </svg>
-              ) : (
-                <svg className="w-5 h-5 text-gray-700" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z" />
-                </svg>
-              )}
-            </button>
-            {/* Hamburger */}
-            <button
-              className="focus:outline-none p-1.5"
-              onClick={() => setMobileMenuOpen(o => !o)}
-              aria-label={isMobileMenuOpen ? "Close menu" : "Open menu"}
-            >
-              {isMobileMenuOpen ? (
-                <svg className="w-5 h-5" fill="none" stroke="currentColor" strokeWidth={2} viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
-                </svg>
-              ) : (
-                <svg className="w-5 h-5" fill="none" stroke="currentColor" strokeWidth={2} viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" d="M4 6h16M4 12h16M4 18h16" />
-                </svg>
-              )}
-            </button>
-          </div>
-        </div>
-
-        {/* --- Mobile slide‑down menu --- */}
-        <AnimatePresence>
-          {isMobileMenuOpen && (
-            <motion.div
-              className={`sm:hidden fixed top-[52px] left-0 right-0 border-t shadow-lg ${
-                isDarkMode ? "bg-gray-800 border-gray-700" : "bg-white border-orange-500"
-              }`}
-              initial={{ height: 0, opacity: 0 }}
-              animate={{ height: "auto", opacity: 1 }}
-              exit={{ height: 0, opacity: 0 }}
-              transition={{ duration: 0.2 }}
-            >
-              {navLinks.map(link => (
-                <a
-                  key={link.id}
-                  href={`#${link.id}`}
-                  onClick={() => {
-                    setMobileMenuOpen(false);
-                    document.getElementById(link.id)?.scrollIntoView({ behavior: "smooth" });
-                  }}
-                  className={`block px-4 py-3 border-b transition-colors duration-200 ${
-                    isDarkMode ? "border-gray-700 hover:bg-gray-700" : "border-gray-200 hover:bg-gray-100"
-                  }`}
-                >
-                  {link.label}
-                </a>
-              ))}
-            </motion.div>
-          )}
-        </AnimatePresence>
-      </nav>
+      <Navbar isDarkMode={isDarkMode} setIsDarkMode={setIsDarkMode} />
 
       {/* HOME SECTION */}
       <section id="home" className="min-h-screen flex flex-col lg:flex-row relative pt-16">
@@ -1484,16 +1304,7 @@ function App() {
       </section>
 
       {/* FOOTER */}
-      <footer className={`${isDarkMode ? 'bg-gray-900/90' : 'bg-gray-900/90'} text-white py-6 backdrop-blur-sm`}>
-                    <motion.div 
-                      initial={{ opacity: 0, y: 20 }}
-                      whileInView={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.6 }}
-          className="text-center text-sm"
-        >
-          © {new Date().getFullYear()} Enable Intelligence. All rights reserved.
-        </motion.div>
-      </footer>
+      <Footer isDarkMode={isDarkMode} />
 
       <AnimatePresence mode="wait">
         {isLoading ? (
